+++ crates/cuid1/src/text.rs
@@ -1,40 +1,10 @@
+use std::iter;
 use std::{char, cmp::Ordering};
-use std::{f64, iter};
 
 use crate::error::CuidError;
-use crate::BASE;
-
-/// Convert a number to a string of the specified radix.
-fn to_radix_string<N: Into<u128>>(radix: u8, number: N) -> Result<String, CuidError> {
-    // Ok(base36::encode(&number.into().to_le_bytes()))
-    let mut number = number.into();
-    let rad_u32: u32 = radix.into();
-
-    if number < radix.into() {
-        // No need to allocate a vector or do any math
-        // NOTE: we are okay to cast to u32 here, b/c number < radix,
-        // which has to be 255 or below.
-        return char::from_digit(number as u32, rad_u32)
-            .map(|c| c.to_string())
-            .ok_or(CuidError::TextError("Bad digit"));
-    } else if number > f64::MAX as u128 {
-        return Err(CuidError::TextError("Input number too large"));
-    }
-
-    // 32 chars should almost always be enough to fill without needing to grow
-    let mut chars: Vec<char> = Vec::with_capacity(32);
-    while number > 0 {
-        // We can unwrap here b/c we know that the modulus must be less than the
-        // radix, which is less than 256
-        chars.push(char::from_digit((number % radix as u128) as u32, rad_u32).unwrap());
-        number /= radix as u128;
-    }
-    chars.reverse();
-    Ok(chars.into_iter().collect())
-}
 
 pub fn to_base_string<N: Into<u128>>(number: N) -> Result<String, CuidError> {
-    to_radix_string(BASE, number)
+    Ok(cuid_util::to_base_36(number))
 }
 
 /// Pad a string `to_pad` up to size `size` with char `char`
@@ -95,36 +65,6 @@ mod pad_tests {
     #[test]
     fn pad_0s() {
         assert_eq!("00foo", &*pad(5, "foo".into()))
-    }
-}
-
-#[cfg(test)]
-mod radix_str_tests {
-    use super::*;
-
-    #[test]
-    fn hex_number_below_radix() {
-        assert_eq!("8", &*to_radix_string(16, 8u8).unwrap());
-    }
-
-    #[test]
-    fn hex_number_below_radix_letter() {
-        assert_eq!("a", &*to_radix_string(16, 10u8).unwrap());
-    }
-
-    #[test]
-    fn number_above_radix() {
-        assert_eq!("10", &*to_radix_string(16, 16u8).unwrap())
-    }
-
-    #[test]
-    fn number_well_above_radix() {
-        assert_eq!("16i", &*to_radix_string(32, 1234u16).unwrap())
-    }
-
-    #[test]
-    fn large_base_36() {
-        assert_eq!("7cik2", &*to_radix_string(36, 12341234u32).unwrap())
     }
 }